add option to use QE bands for gap values in gap.py
To check if Wannier functions are breaking 3-fold symmetry of d,
check the bands at K from QE instead of those obtained by diagonalizing
the Wannier Hamiltonian. The eigenstates from the Wannier Hamiltonian
are still used for layer composition determination.

To check this symmetry at the appropriate level of precision using the
QE bands, needed to increase the hard-coded precision at which QE's
bands.x outputs eigenvalues. This required a change in extractQEBands.py
to use this modified format.